[BE]Changed model variables int to double for accuracy

+++ ThucDon/Models/CarbFood.cs
@@ -17,13 +17,13 @@ namespace ThucDon.Models
         [Display(Name = "Food Name")]
         public string Name { get; set; }
         [Required]
-        public int Protein { get; set; }
+        public double Protein { get; set; }
         [Required]
-        public int Fat { get; set; }
+        public double Fat { get; set; }
         [Required]
-        public int Carb { get; set; }
+        public double Carb { get; set; }
         [Required]
-        public int Fiber { get; set; }
+        public double Fiber { get; set; }
 
         [Display(Name = "Food Category")]
         [ForeignKey("FoodCategoryID")]
